hrms-lite-backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.2
django-cors-headers==4.3.1

# PostgreSQL driver (Psycopg2 binary build)
psycopg2-binary==2.9.11

# WSGI server for production
gunicorn==23.0.0

# Optional: environment variables helper (if you want it)
python-dotenv==1.0.1
